// utils/types.ts
  export type Coin = {
    item: Item;
  }
  interface Item {
    id: string;
    coin_id: number;
    name: string;
    symbol: string;
    market_cap_rank: number;
    thumb: string;
    small: string;
    large: string;
    slug: string;
    price_btc: number;
    score: number;
    data: Data;
  }
  interface Data {
    price: string;
    price_btc: string;
    price_change_percentage_24h: Pricechangepercentage24h;
    market_cap: string;
    market_cap_btc: string;
    total_volume: string;
    total_volume_btc: string;
    sparkline: string;
    content?: any;
  }